chore: refactor setup tls server
Refactors the tls server setup from `test/setup-tls-server.ts` into a
reusable test helper module.
